requirements.txt
requests
Gooey
PyExecJS
pyinstaller